% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{clusters_per_read}
\alias{clusters_per_read}
\title{Obtain a series of tandem repeat unit clusters per sequencing read}
\usage{
clusters_per_read(tsclust_obj)
}
\arguments{
\item{tsclust_obj}{A tsclust object generated by dtwclust::tsclust()}
}
\description{
clusters_per_read() extracts the cluster per tandem repeat unit and 
concatenates them according to their original sequence in the sequencing
read.
}
\details{
This function uses a tsclust object generated by dtwclust::tsclust() and
returns a data.frame containing a comma-separated series of tandem repeat
unit clusters.
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/")
qc <- summary_qc(df)
df2 <- qual_reads(df,qc$center_cutoff)
squiggles <- load_squiggles("/storage/NanoSatellite_chunks/",df2)

library(doParallel)
library(dtwclust)
k_clusters=2
registerDoParallel(cores=8)

positive_clustering=tsclust(squiggles$positive,type="h",k=k_clusters,trace=TRUE,distance = "dtw_basic", control=hierarchical_control(method="ward.D",symmetric = T))

cpr=clusters_per_read(positive_clustering)
}
