% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{extract_centroids}
\alias{extract_centroids}
\title{Extract tandem repeat unit squiggle centroids.}
\usage{
extract_centroids(tsclust_obj)
}
\arguments{
\item{tsclust_obj}{A tsclust object generated by dtwclust::tsclust()}
}
\description{
extract_centroids() obtains the squiggle current levels which best represent
their respective clusters.
}
\details{
dtwclust::tsclust() can be used to cluster tandem repeat unit squiggles
and centroids are calculated for each cluster. This centroid
aims to provide the the most representative squiggle current level for 
each cluster. extract_centroids extracts the centroid information from a
tsclust object and returns a data.frame with the results
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/")
qc <- summary_qc(df)
df2 <- qual_reads(df,qc$center_cutoff)
squiggles <- load_squiggles("/storage/NanoSatellite_chunks/",df2)

library(doParallel)
library(dtwclust)
k_clusters=2
registerDoParallel(cores=8)

positive_clustering=tsclust(squiggles$positive,type="h",k=k_clusters,trace=TRUE,distance = "dtw_basic", control=hierarchical_control(method="ward.D",symmetric = T))

cent=extract_centroids(positive_clustering)
ggplot(cent,aes(x=pos,y=signal,colour=factor(cluster)))+geom_point()+geom_line()+theme_minimal()+facet_grid(. ~ cluster)+guides(colour=guide_legend(title="cluster"))
}
