% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{heatmap_custom}
\alias{heatmap_custom}
\title{Custom heatmap function}
\usage{
heatmap_custom(x, Rowv = NULL, Colv = if (symm) "Rowv" else NULL,
  distfun = dist, hclustfun = hclust, reorderfun = function(d, w)
  reorder(d, w), add.expr, symm = FALSE, revC = identical(Colv,
  "Rowv"), scale = c("row", "column", "none"), na.rm = TRUE,
  margins = c(5, 5), ColSideColors, RowSideColors, cexRow = 0.2 +
  1/log10(nr), cexCol = 0.2 + 1/log10(nc), labRow = NULL,
  labCol = NULL, main = NULL, xlab = NULL, ylab = NULL,
  keep.dendro = FALSE, verbose = getOption("verbose"), lwd = 10, ...)
}
\description{
This function is slightly adapted from the base heatmap() function
to allow control over the line width of the dendrograms. It is
recommended to run ns_heatmap() which incorporates heatmap_custom().
}
