% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{load_summary}
\alias{load_summary}
\title{Loads summary files generated by Signal2chunk.R}
\usage{
load_summary(dir)
}
\arguments{
\item{dir}{Path denoted as a character vector of length 1.}
}
\description{
load_summary() returns a data frame which combines summary files generated 
by Signal2chunk.R
}
\details{
This function requires a path to the results produced by Signal2chunk.R 
load_summary() will recursively search the directory and combine the 
.table files in a single dataframe.
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/") 
}
