% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Wrapper function to plot tandem repeat lengths for multiple samples.}
\usage{
plot_lengths(df, binwidth = 5)
}
\arguments{
\item{df}{data.frame generated by load_summary(), or qual_reads()}

\item{binwidth}{A numeric value passed on to geom_dotplot()}
}
\description{
plot_lengths() returns a ggplot2 based plot with each sample in a
panel, the number of tandem repeat units on the y-axis, and colored
dots corresponding to positive (red) and negative (blue) DNA strands
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/")
qc <- summary_qc(df)
df2 <- qual_reads(df,qc$center_cutoff)
plot_lengths(df2)
}
