% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{qual_reads}
\alias{qual_reads}
\title{Wrapper function to filter sequencing reads based on distance cutoffs}
\usage{
qual_reads(df, center_cutoff = 1, flank_cutoff = 1)
}
\arguments{
\item{df}{A data.frame, generated by load_summary()}

\item{center_cutoff}{A numeric value [0-1] corresponding to the maximum
center normalized dynamic time warping distance}

\item{flank_cutoff}{A numeric value [0-1] corresponding to the maximum
flank normalized dynamic time warping distance}
}
\description{
Returns a data.frame containing sequencing reads with a "flank" and/or 
"center" normalized dynamic time warping distance smaller or equal to
a user specified cutoff value.
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/")
qc <- summary_qc(df)
df2 <- qual_reads(df,qc$center_cutoff)
}
