% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoSatelliteR_functions.R
\name{summary_qc}
\alias{summary_qc}
\title{Quality control of NanoSatellite delineation and segmentation}
\usage{
summary_qc(df)
}
\arguments{
\item{df}{data.frame generated by load_summary()}
}
\description{
summary_qc() returns (box)plots and suggested cutoff values to filter 
tandem repeat segmentations based on normalized distance.
}
\details{
This function requires a data frame as generated by load_summary().
summary_qc() will analyze the normalized "flank" and  
"center" dynamic time warping distance,
respectively corresponding to delineation of tandem repeat squiggles
from flanking squiggles, and the segmentation of the tandem repeat
squiggle. A scatter plot of both distances is shown as well as boxplots
for the individual metrics. In addition, cutoff values corresponding 
to 1.5 times the interquartile range from 
the 75th percentile are returned.
}
\examples{
df <- load_summary("/storage/NanoSatellite_chunks/")
qc <- summary_qc(df)
}
